refactor(presets): added models module for presets and changed code accordingly
Added a new models module for all preset related dataclasses/models. The Preset class from
the preset_manager was moved here and the code was (almost) changed accordingly.

BREAKING CHANGE: Code not changed completely
